CV website HTML, CSS & Javascript
I present a CV website for myself that I coded but never hosted. This
was done as a project for my Web & UI design module. I have uploaded all
the files needed to run this website properly - including the colorbox
CSS and JS files which were not written by myself.

What I have written are four HTML files and a CSS file. Many image files
are included and these are needed to run the website.